Allow Authn to bypass nil users, given an option
There are cases, particularly API calls that can be done both when
authenticated and not, where we actually may want to allow a `nil`
current user.

This is currently not possible because `Authn` always requires a user to
be set.

This proposal adds a `require: boolean` option to `Authn` so that, if
the user is `nil`, it still sets the Speakeasy context just the same.
Defaults to `true` (the current behaviour).

Fixes #12.

Example usage:

```elixir
middleware(Speakeasy.Authn, require: false)
```